requirements.txt
flask==2.3.3
discord.py==2.3.2
aiohttp==3.8.5
python-dotenv==1.0.0
